Make tv_sec a time_t on all platforms but alpha.  Brings us more in line with POSIX.  This also makes the struct correct we ever implement an i386-time64 architecture.  Not that we need too.
Reviewed by:	imp, brooks
Approved by:	njl (acpica), des (no objects, touches procfs)
Tested with:	make universe


git-svn-id: svn+ssh://svn.freebsd.org/base/head@153706 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/dev/acpica/Osd/OsdSynch.c
@@ -257,8 +257,8 @@ AcpiOsWaitSemaphore(ACPI_HANDLE Handle, UINT32 Units, UINT16 Timeout)
 	    tmo = 1;
 
 	if (acpi_semaphore_debug) {
-	    printf("%s: Wakeup timeleft(%lu, %lu), tmo %u, sem %p, thread %d\n",
-		__func__, timelefttv.tv_sec, timelefttv.tv_usec, tmo, as,
+	    printf("%s: Wakeup timeleft(%jd, %lu), tmo %u, sem %p, thread %d\n",
+		__func__, (intmax_t)timelefttv.tv_sec, timelefttv.tv_usec, tmo, as,
 		AcpiOsGetThreadId());
 	}
     }
